Service class added

+++ src/main/java/com/example/webapp/model/Postcard.java
@@ -4,17 +4,18 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.validation.constraints.NotBlank;
 
 @Entity
 public class Postcard {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     private String postNumber;
     //отправитель/получатель
     //@NotNull
-    //@NotBlank
+//    @NotBlank
     private String country;
 
     public Postcard(){
@@ -26,11 +27,11 @@ public class Postcard {
         this.country = country;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
